Extract firebase code from FileUploadPage
